refac(api): refactor models and controllers
requested changes
