+++ cooking-user/sj2324-5ehif-cooking-user.Application/DTO/LoginModel.cs
@@ -1,0 +1,7 @@
+namespace sj2324_5ehif_cooking_user.Application.DTO;
+
+public record LoginModel
+{
+    public string? Username { get; set; }
+    public string? Password { get; set; }
+}
