+++ cooking-user/sj2324-5ehif-cooking-user.Test/UserTests/AuthTest.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
+using sj2324_5ehif_cooking_user.Application.DTO;
 using sj2324_5ehif_cooking_user.Application.Infrastructure;
 using sj2324_5ehif_cooking_user.Application.Model;
 using sj2324_5ehif_cooking_user.Webapi.Controllers;
